requirements.txt
streamlit
openai
chromadb
pdfplumber
python-docx
tqdm
pandas
python-multipart
requests
sentence-transformers
